Move SierraAccessStatus into the source_model lib
