+++ pipeline/transformer/transformer_sierra/src/main/scala/uk/ac/wellcome/platform/transformer/sierra/transformers/SierraLocation.scala
@@ -1,7 +1,10 @@
 package uk.ac.wellcome.platform.transformer.sierra.transformers
 
 import weco.catalogue.internal_model.locations._
-import weco.catalogue.source_model.sierra.rules.SierraPhysicalLocationType
+import weco.catalogue.source_model.sierra.rules.{
+  SierraAccessStatus,
+  SierraPhysicalLocationType
+}
 import weco.catalogue.source_model.sierra.{
   SierraBibData,
   SierraBibNumber,
